Fixe bug with marking(it's not final version)
