+++ testing_system/test.c
@@ -5,11 +5,13 @@
 typedef struct A { int* a; int* b; }A;
 
 int main() {
-	//A* a = (A*)allocate_new_object(sizeof(A));
-	//A* c = (A*)allocate_new_object(sizeof(A));
-	//A* b = (A*)allocate_new_object(sizeof(A));
-	//printf("%p\n", &a);
-	int* a = allocate_new_object(40);
+	int* a = (int*)allocate_new_object(80);
+	int* b = (int*)allocate_new_object(80);
+	int* c = (int*)allocate_new_object(40);
 	segment_traverse(end_rsp_value, start_rsp_value);
+	
+	//show_bitmap(a);	
+	//printf("START = %llu\nEND = %llu\n%llu\n", START_ALLOCATOR_HEAP, END_ALLOCATOR_HEAP, a);
 	return 0;
+
 }
